+++ include/webserver.h
@@ -21,7 +21,7 @@ void playStationInt(int sid);
 void websockethandle(int socket, wsopcode_t opcode, uint8_t * payload, size_t length);
 uint16_t getVolume(void);
 void setVolume(char* vol);
-void setVolumei(int16_t vol);
+void setVolumei(int8_t vol);
 void setRelVolume(int8_t vol);
 
 #endif
